requirements.txt
Flask
unidecode
rembg
Pillow
numpy
uvicorn
onnxruntime